+++ build.gradle
@@ -7,12 +7,9 @@ allprojects {
         dependencies {
             compileOnly group: 'org.projectlombok', name: 'lombok', version: '1.18.20'
             annotationProcessor 'org.projectlombok:lombok:1.18.20'
-            testImplementation("org.junit.jupiter:junit-jupiter:5.7.0")
-            testImplementation group: 'org.springframework', name: 'spring-test', version: '5.3.8'
-            implementation group: 'org.springframework', name: 'spring-context', version: '5.3.7'
-            implementation group: 'org.springframework', name: 'spring-webmvc', version: '5.3.7'
-            implementation group: 'org.slf4j', name: 'slf4j-api', version: '2.0.0-alpha1'
-            implementation group: 'org.slf4j', name: 'slf4j-simple', version: '2.0.0-alpha1'
+            implementation group: 'org.springframework.boot', name: 'spring-boot-starter-web', version: '2.5.1'
+            testImplementation group: 'org.springframework.boot', name: 'spring-boot-starter-test', version: '2.5.1'
+            implementation group: 'org.springframework.boot', name: 'spring-boot-starter-validation', version: '2.5.1'
         }
 
         test {
