+++ controller/src/main/java/com/epam/esm/config/ApplicationConfiguration.java
@@ -1,0 +1,25 @@
+package com.epam.esm.config;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.MessageSource;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.support.ResourceBundleMessageSource;
+
+@SpringBootApplication
+@ComponentScan("com.epam.esm")
+public class ApplicationConfiguration {
+    public static void main(String[] args) {
+        SpringApplication.run(ApplicationConfiguration.class, args);
+    }
+
+    @Bean
+    public MessageSource messageSource() {
+        ResourceBundleMessageSource source = new ResourceBundleMessageSource();
+        source.setBasename("property/message");
+        source.setUseCodeAsDefaultMessage(true);
+        source.setDefaultEncoding("UTF-8");
+        return source;
+    }
+}
