--- controller/src/main/java/com/epam/esm/config/ConfigWebApplicationInitializer.java
@@ -1,17 +1,0 @@
-package com.epam.esm.config;
-
-import org.springframework.web.WebApplicationInitializer;
-
-import javax.servlet.ServletContext;
-import javax.servlet.ServletException;
-
-/**
- * The type Config web application initializer.
- */
-public class ConfigWebApplicationInitializer implements WebApplicationInitializer {
-
-    @Override
-    public void onStartup(ServletContext servletContext) throws ServletException {
-        servletContext.setInitParameter("spring.profiles.active", "prod");
-    }
-}
