+++ model/build.gradle
@@ -1,3 +1,3 @@
 plugins {
-    id 'war'
+    id 'java'
 }
