+++ persistence/build.gradle
@@ -1,13 +1,10 @@
 plugins {
-    id 'war'
+    id 'java'
 }
 
 dependencies {
     implementation project(":model")
-    implementation group: 'com.zaxxer', name: 'HikariCP', version: '4.0.3'
     implementation group: 'mysql', name: 'mysql-connector-java', version: '8.0.25'
-    implementation group: 'org.springframework', name: 'spring-jdbc', version: '5.3.7'
-    testImplementation group: 'org.mockito', name: 'mockito-junit-jupiter', version: '3.6.28'
+    implementation group: 'org.springframework.boot', name: 'spring-boot-starter-jdbc', version: '2.5.1'
     testImplementation group: 'com.h2database', name: 'h2', version: '1.4.200'
-
 }
