+++ persistence/src/main/java/com/epam/esm/config/DatabaseConfig.java
@@ -2,12 +2,10 @@ package com.epam.esm.config;
 
 import com.zaxxer.hikari.HikariConfig;
 import com.zaxxer.hikari.HikariDataSource;
-import lombok.AllArgsConstructor;
+import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Profile;
-import org.springframework.context.annotation.PropertySource;
-import org.springframework.core.env.Environment;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 
@@ -17,11 +15,21 @@ import javax.sql.DataSource;
  * The type Database config.
  */
 @Configuration
-@PropertySource("classpath:property/dataSource-prod.properties")
-@AllArgsConstructor
+@RequiredArgsConstructor
 public class DatabaseConfig {
 
-    private final Environment environment;
+
+    @Value("${spring.datasource.username}")
+    private String jdbcName;
+
+    @Value("${spring.datasource.url}")
+    private String jdbcUrl;
+
+    @Value("${spring.datasource.password}")
+    private String jdbcPassword;
+
+    @Value("${spring.datasource.driver-class-name}")
+    private String jdbcDriver;
 
     /**
      * Gets data source.
@@ -29,14 +37,12 @@ public class DatabaseConfig {
      * @return the data source
      */
     @Bean
-    @Profile("prod")
     public DataSource getDataSource() {
         HikariConfig config = new HikariConfig();
-        config.setJdbcUrl(environment.getProperty("url"));
-        config.setUsername(environment.getProperty("user"));
-        config.setPassword(environment.getProperty("password"));
-        config.setDriverClassName(environment.getProperty("driverClassName"));
-        config.setMaximumPoolSize(Integer.parseInt(environment.getProperty("maximum-pool-size")));
+        config.setJdbcUrl(jdbcUrl);
+        config.setUsername(jdbcName);
+        config.setPassword(jdbcPassword);
+        config.setDriverClassName(jdbcDriver);
         return new HikariDataSource(config);
     }
