+++ persistence/src/main/resources/application-prod.properties
@@ -1,0 +1,4 @@
+spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
+spring.datasource.url=jdbc:mysql://localhost:3306/certificate
+spring.datasource.username=root
+spring.datasource.password=2745839
