--- persistence/src/main/resources/property/dataSource-prod.properties
@@ -1,15 +1,0 @@
-url=jdbc:mysql://localhost:3306/certificate
-driverClassName=com.mysql.cj.jdbc.Driver
-user=root
-password=2745839
-characterEncoding=UTF-8
-useUnicode=true
-serverTimezone=UTC
-useSSL=false
-requireSSL=false
-allowPublicKeyRetrieval=true
-cachePrepStmts = true
-connection-timeout=10000
-maximum-pool-size=10
-minimum-idle=15
-pool-name=HikariConnectionPool
