+++ persistence/src/test/resources/application-dev.properties
@@ -1,0 +1,4 @@
+spring.datasource.url=jdbc:h2:mem:testdb;database_to_lower=true;CASE_INSENSITIVE_IDENTIFIERS=TRUE;INIT=RUNSCRIPT FROM 'classpath:create_schema.sql'
+spring.datasource.driver-class-name=org.h2.Driver
+spring.datasource.username=admin
+spring.datasource.password=
