--- persistence/src/test/resources/property/dataSource-dev.properties
@@ -1,5 +1,0 @@
-datasource.url=jdbc:h2:mem:testdb;database_to_lower=true;CASE_INSENSITIVE_IDENTIFIERS=TRUE;INIT=RUNSCRIPT FROM 'classpath:create_schema.sql'
-datasource.driverClassName=org.h2.Driver
-datasource.username=admin
-datasource.password=
-jpa.database-platform=org.hibernate.dialect.H2Dialect
